* some clustering work * fixed damage shield crashes
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@1723 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/Spell.cpp
@@ -2883,7 +2883,7 @@ int8 Spell::CheckItems()
 				if( proto->DisenchantReqSkill < 1 )
 					return SPELL_FAILED_CANT_BE_DISENCHANTED;
 
-				if( p_caster->_GetSkillLineCurrent(SKILL_ENCHANTING) < proto->DisenchantReqSkill )
+				if( (int32)p_caster->_GetSkillLineCurrent(SKILL_ENCHANTING) < proto->DisenchantReqSkill )
 					return SPELL_FAILED_CANT_BE_DISENCHANTED_SKILL;
 			}
 			else
